refactor: reorganize "usecase" directory structure; updated imports and tests accordingly
