+++ llm_chat/domain/use_case/completion/base.py
