+++ llm_chat/domain/use_case/completion/rag.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import User
 
 from lib.llm.valueobject.completion import RoleType
 from llm_chat.domain.service.completion.rag import OpenAIRagChatService
-from llm_chat.domain.usecase.completion.base import UseCase
+from llm_chat.domain.use_case.completion.base import UseCase
 from llm_chat.domain.valueobject.completion.chat import MessageDTO
 from llm_chat.domain.valueobject.completion.use_case import UseCaseType
